WIP attempting to fix dtype cannot be object error

+++ orpheus_tts/engine_class.py
@@ -64,18 +64,22 @@ class OrpheusModel:
                 adapted_prompt = f"{voice}: {prompt}"
                 prompt_tokens = self.tokeniser(
                     adapted_prompt, return_tensors="pt")
-                start_token = torch.tensor([[128259]], dtype=torch.int64)
+                # start_token = torch.tensor([[128259]], dtype=torch.int64)
+                start_token = torch.tensor([[128259]])
                 end_tokens = torch.tensor(
-                    [[128009, 128260, 128261, 128257]], dtype=torch.int64)
+                    # [[128009, 128260, 128261, 128257]], dtype=torch.int64)
+                    [[128009, 128260, 128261, 128257]])
                 all_input_ids = torch.cat(
                     [start_token, prompt_tokens.input_ids, end_tokens], dim=1)
                 prompt_string = self.tokeniser.decode(all_input_ids[0])
                 return prompt_string
             else:
                 prompt_tokens = self.tokeniser(prompt, return_tensors="pt")
-                start_token = torch.tensor([[128259]], dtype=torch.int64)
+                # start_token = torch.tensor([[128259]], dtype=torch.int64)
+                start_token = torch.tensor([[128259]])
                 end_tokens = torch.tensor(
-                    [[128009, 128260, 128261, 128257]], dtype=torch.int64)
+                    # [[128009, 128260, 128261, 128257]], dtype=torch.int64)
+                    [[128009, 128260, 128261, 128257]])
                 all_input_ids = torch.cat(
                     [start_token, prompt_tokens.input_ids, end_tokens], dim=1)
                 prompt_string = self.tokeniser.decode(all_input_ids[0])
